+++ testing/adios2/engine/ssc/TestSscXgc3Way.cpp
@@ -25,8 +25,9 @@ public:
 void coupler(const Dims &shape, const Dims &start, const Dims &count,
              const size_t steps, const adios2::Params &engineParams)
 {
-    size_t datasize = std::accumulate(count.begin(), count.end(), 1,
-                                      std::multiplies<size_t>());
+    size_t datasize =
+        std::accumulate(count.begin(), count.end(), static_cast<size_t>(1),
+                        std::multiplies<size_t>());
 
     adios2::ADIOS adios(mpiComm);
 
@@ -76,7 +77,8 @@ void coupler(const Dims &shape, const Dims &start, const Dims &count,
         auto x_to_c_var = x_to_c_io.InquireVariable<float>("x_to_c");
         auto readShape = x_to_c_var.Shape();
         x_to_c_data.resize(std::accumulate(readShape.begin(), readShape.end(),
-                                           1, std::multiplies<size_t>()));
+                                           static_cast<size_t>(1),
+                                           std::multiplies<size_t>()));
         x_to_c_engine.Get(x_to_c_var, x_to_c_data.data(), adios2::Mode::Sync);
         VerifyData(x_to_c_data.data(), i, Dims(readShape.size(), 0), readShape,
                    readShape, mpiRank);
@@ -91,7 +93,8 @@ void coupler(const Dims &shape, const Dims &start, const Dims &count,
         auto g_to_c_var = g_to_c_io.InquireVariable<float>("g_to_c");
         readShape = g_to_c_var.Shape();
         g_to_c_data.resize(std::accumulate(readShape.begin(), readShape.end(),
-                                           1, std::multiplies<size_t>()));
+                                           static_cast<size_t>(1),
+                                           std::multiplies<size_t>()));
         g_to_c_engine.Get(g_to_c_var, g_to_c_data.data(), adios2::Mode::Sync);
         VerifyData(g_to_c_data.data(), i, Dims(readShape.size(), 0), readShape,
                    readShape, mpiRank);
@@ -114,8 +117,9 @@ void coupler(const Dims &shape, const Dims &start, const Dims &count,
 void xgc(const Dims &shape, const Dims &start, const Dims &count,
          const size_t steps, const adios2::Params &engineParams)
 {
-    size_t datasize = std::accumulate(count.begin(), count.end(), 1,
-                                      std::multiplies<size_t>());
+    size_t datasize =
+        std::accumulate(count.begin(), count.end(), static_cast<size_t>(1),
+                        std::multiplies<size_t>());
 
     std::vector<float> x_to_c_data(datasize);
     std::vector<float> c_to_x_data;
@@ -151,7 +155,8 @@ void xgc(const Dims &shape, const Dims &start, const Dims &count,
         auto c_to_x_var = c_to_x_io.InquireVariable<float>("c_to_x");
         auto readShape = c_to_x_var.Shape();
         c_to_x_data.resize(std::accumulate(readShape.begin(), readShape.end(),
-                                           1, std::multiplies<size_t>()));
+                                           static_cast<size_t>(1),
+                                           std::multiplies<size_t>()));
         c_to_x_engine.Get(c_to_x_var, c_to_x_data.data(), adios2::Mode::Sync);
         VerifyData(c_to_x_data.data(), i, Dims(readShape.size(), 0), readShape,
                    readShape, mpiRank);
@@ -185,8 +190,9 @@ void gene(const Dims &shape, const Dims &start, const Dims &count,
         g_to_c_io.Open("g_to_c", adios2::Mode::Write);
     g_to_c_engine.LockWriterDefinitions();
 
-    size_t datasize = std::accumulate(shape.begin(), shape.end(), 1,
-                                      std::multiplies<size_t>());
+    size_t datasize =
+        std::accumulate(shape.begin(), shape.end(), static_cast<size_t>(1),
+                        std::multiplies<size_t>());
     std::vector<float> c_to_g_data;
     std::vector<float> g_to_c_data(datasize);
 
@@ -196,7 +202,8 @@ void gene(const Dims &shape, const Dims &start, const Dims &count,
         auto c_to_g_var = c_to_g_io.InquireVariable<float>("c_to_g");
         auto readShape = c_to_g_var.Shape();
         c_to_g_data.resize(std::accumulate(readShape.begin(), readShape.end(),
-                                           1, std::multiplies<size_t>()));
+                                           static_cast<size_t>(1),
+                                           std::multiplies<size_t>()));
         c_to_g_engine.Get(c_to_g_var, c_to_g_data.data(), adios2::Mode::Sync);
         VerifyData(c_to_g_data.data(), i, Dims(readShape.size(), 0), readShape,
                    readShape, mpiRank);
